# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem pelaporan keuangan dengan manajemen users, transaksi, dan master data",
        "title": "Pelaporan Keuangan API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/transaksi": {
            "get": {
                "description": "Get all transactions with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaksi"
                ],
                "summary": "Get all transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all transactions success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseGetAllSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dtos.ResTransaksi"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/definitions/helpers.Pagination"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid pagination data",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "404": {
                        "description": "No transactions found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new financial transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaksi"
                ],
                "summary": "Create new transaction",
                "parameters": [
                    {
                        "description": "Transaction data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.InputTransaksi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create transaction success",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseCUDSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid input data or validation error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            }
        },
        "/transaksi/{id}": {
            "get": {
                "description": "Get transaction details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaksi"
                ],
                "summary": "Get transaction details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get transaction detail success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseGetDetailSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dtos.ResTransaksi"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid transaction ID",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing transaction with partial data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaksi"
                ],
                "summary": "Update transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update transaction data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateTransaksiRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update transaction success",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseCUDSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid transaction ID or request data",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a specific transaction by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaksi"
                ],
                "summary": "Delete transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete transaction success",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseCUDSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid transaction ID",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            }
        },
        "/transaksi/{id}/status": {
            "patch": {
                "description": "Update the status of a specific transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaksi"
                ],
                "summary": "Update transaction status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update transaction status success",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseCUDSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid transaction ID or missing status_id",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dtos.InputTransaksi": {
            "type": "object",
            "required": [
                "id_jenis_pembayaran",
                "id_kategori",
                "id_tipe_transaksi",
                "jumlah",
                "nama_transaksi",
                "tanggal"
            ],
            "properties": {
                "bukti_transaksi": {
                    "description": "atau path lokal",
                    "type": "string"
                },
                "id_jenis_pembayaran": {
                    "type": "integer"
                },
                "id_kategori": {
                    "type": "integer"
                },
                "id_tipe_transaksi": {
                    "type": "integer"
                },
                "jumlah": {
                    "type": "number"
                },
                "keterangan": {
                    "type": "string"
                },
                "nama_transaksi": {
                    "type": "string",
                    "maxLength": 255
                },
                "tanggal": {
                    "description": "yyyy-mm-dd",
                    "type": "string"
                }
            }
        },
...
        "dtos.ResTransaksi": {
            "type": "object",
            "properties": {
                "bukti_transaksi": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_jenis_pembayaran": {
                    "type": "integer"
                },
                "id_kategori": {
                    "type": "integer"
                },
                "id_status_transaksi": {
                    "type": "integer"
                },
                "id_tipe_transaksi": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "jumlah": {
                    "type": "number"
                },
                "keterangan": {
                    "type": "string"
                },
                "komentar_manajer": {
                    "type": "string"
                },
                "tanggal": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dtos.UpdateTransaksiRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "bukti_transaksi": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_jenis_pembayaran": {
                    "type": "integer"
                },
                "id_kategori": {
                    "type": "integer"
                },
                "id_status_transaksi": {
                    "type": "integer"
                },
                "id_tipe_transaksi": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "jumlah": {
                    "type": "number"
                },
                "keterangan": {
                    "type": "string"
                },
                "komentar_manajer": {
                    "type": "string"
                },
                "tanggal": {
                    "type": "string"
                }
            }
        },
        "helpers.Pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "example": 1
                },
                "next_page": {
                    "type": "integer",
                    "example": 2
                },
                "page_size": {
                    "type": "integer",
                    "example": 5
                },
                "previous_page": {
                    "type": "integer",
                    "example": 0
                },
                "total_data": {
                    "type": "integer",
                    "example": 50
                },
                "total_page": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "helpers.ResponseCUDSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "helpers.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error message"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "helpers.ResponseGetAllSuccess": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success message"
                },
                "pagination": {
                    "$ref": "#/definitions/helpers.Pagination"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "helpers.ResponseGetDetailSuccess": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success message"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}